requirements.txt
vkbottle
pygsheets
python-dotenv